+++ src/Utils/CurveMode.cs
@@ -1,0 +1,6 @@
+﻿namespace Celeste.Mod.CommunalHelper.Utils {
+    public enum CurveType {
+        Quadratic = 2,
+        Cubic = 3,
+    }
+}
